src/2. BigData/5. Experiments/pyproject.toml
[project]
name = "experiments-canary"
version = "0.1.0"
description = "Canary deployment demo: Serious vs Funny chatbot versions"
requires-python = ">=3.10,<3.13"
dependencies = [
    "openai>=1.0.0",
    "streamlit>=1.25.0",
    "python-dotenv>=1.0.0",
]
